Rename datalab_cohorts to cohortextractor
