+++ MANIFEST.in
@@ -1,2 +1,2 @@
 include runner/VERSION
-include datalab_cohorts/uk_population_bands_2018.csv
+include cohortextractor/uk_population_bands_2018.csv
