+++ setup.py
@@ -19,7 +19,7 @@ setup(
         "sqlparse",
     ],
     entry_points={
-        "console_scripts": ["cohortextractor=datalab_cohorts.cohortextractor:main"]
+        "console_scripts": ["cohortextractor=cohortextractor.cohortextractor:main"]
     },
     include_package_data=True,
     classifiers=["License :: OSI Approved :: GNU General Public License v3 (GPLv3)"],
